Moved date parsing to StringUtil
